requirements.txt
# API
fastapi~=0.124.4
uvicorn~=0.38.0
pydantic~=2.12.5
python-multipart~=0.0.20
jinja2~=3.1.6

# RAG Core
sentence-transformers~=5.2.0
langchain-text-splitters~=1.1.0
langchain-community~=0.4.1
pypdf~=6.4.2

# Utils
python-dotenv~=1.2.1
requests~=2.32.5

# math
numpy==2.3.5

# quality
ruff~=0.14.9